Update patch set 1
Patch Set 1: -Code-Review

Well - IMHO it doesnt make sense for this to become an endless discussion. For me this is still "nice to have" and optional cause local_manifest is exactly made for such things. Its not only about adding repos but also to add remotes if you need them. But I will remove my -1 to go on with this either way.

Patch-set: 1