Update patch set 1
Patch Set 1:

Andre - it's more the principle I guess. If we add this remote then why not add device repos into this file as well? Or how about kernel repos and all device dependencies? I'm not for or against this commit but I get why there is an issue over this commit.

Patch-set: 1