Update patch set 1
Patch Set 1:

I don't see what's wrong with how this is evaluted - we're having a healthy discussion, and no one person is forcing it through.  That is the whole point of what we're trying to change.

That said, this might be one of those cases where eventually someone is going to have to step in and make a final decision, evaluating all input.

I'm not going to do that yet though.  :)

My perspective is - It makes things easier for people and has little to no negative impact.  Also, people have mentioned roomservice, so the final fate of this patch might be decided when we get roomservice up and running, as this might be a dependency of roomservice.

Patch-set: 1