Update patch set 1
Patch Set 1: Verified+1

Key here is:  The negatives of having it in here are minimal, and pretty much all other projects do it.

The benefit is instead of it being replicated in multiple local manifests (this is going to be a fairly frequently used remote during device bringup efforts), it's in a central place.

Patch-set: 1
Reviewer: Gerrit User 1000008 <1000008@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Verified=+1