Update patch set 1
Patch Set 1: Code-Review+1

No on removing the hardware/ repos - Removal of repos must be done EXTREMELY carefully.

Repos that are low activity, or in this case, just dead/obsolete are easy.

Repos with active work going on are MUCH harder.  Problem is that you can't cherry-pick manifest changes, they MUST be merged.  Also, until a device repo's dependencies get synced, if a repo is removed from the manifest, it gets removed from people's machines when the repo syncs.  We had some people lose a LOT of work last time we tried to do that.

Problem is that adding stuff to omni.dependencies in advance doesn't work - roomservice detects that the stuff is already there, and forcing duplication in the local manifests causes sync failures.

Patch-set: 1
Label: Code-Review=+1