Update patch set 1
Patch Set 1:

_artem, I will try to read this comment as more of a suggestion than trying to tell people what to do, even though it came across quite demanding.

That feature certainly sounds useful, I don't think anyone is denying it is useful. But bear in mind that you are more than welcome to submit this change, and it would be merged, meaning you had it available. :)

Omni isn't anybody's job here, so unfortunately they are not obliged to introduce a feature since it was requested. But sometimes they will be able to introduce things when they see something of interest. Equally though, if someone else sees this and feels they want it, and implement it, they are able to submit it and have it merged :)

I am sure people can go and install an app if they want. The issue with that is that removal will affect everyone who already uses it, as they will lose their settings if it is removed. That's a valid concern that needs to be considered when removing things.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Verified=0