Update patch set 3
Patch Set 3:

If a lot of the work was done by you, I'm more willing to pull it in as-is.

Also, the negative impact of these devices on existing omni devices should be minimal - I'm not sure if we have any currently supported devices using TI WLAN, so major changes here SHOULD be safer than elsewhere.

Patch-set: 3
Label: Code-Review=0