Update patch set 4
Patch Set 4:

I know that. I have already pushed a series of patches for legacy LPA (topic legacy-lpa) but it seems that it's totally broken in kitkat.

My problem is that the video decoder crashes the device when you try to play anything and causes hard reboots. The problem looks like it's kernel related, but the exact same kernel works in CM without issues :/

I spent like 2 weeks without finding any fix for this

Patch-set: 4