default: Add actual branch for gerrit and repo init
This is just a small change but will save time for new users so
they just need to copy and paste the line in case of the repo
init command.

For the gerrit one just will save some time too, so they just
need to add the project and them username.

Change-Id: I91e2899a9bdcc38fac1c77fedb0e80e82e52f868