Update patch set 5
Patch Set 5:

I agree Andrew that git cherry-pick preserves authorship but there are cases where someone might manually move code to avoid conflicts and/or trying to adopt it to a certain base.

In cases where the base is also completely different, Omni and CM for example, this would be necessary to avoid having to push 10-15 patchsets to gerrit.

In the end, is all a precaution for those that mean well and are trying to do the right thing. If people decide they're not going to maintain authorship, this or anything else won't stop them unfortunately

Patch-set: 5
Reviewer: Gerrit User 1000283 <1000283@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0