Update patch set 5
Patch Set 5:

What I mean Max is by looking at a commit say you committed to Omni and I want to push it to DU. I would look at it and adapt it to DU instead of git cherry-pick.

Opening up a terminal and opening file by file and injecting the code. Once compiled and tested and all, commit it and change the author to you before pushing to gerrit or github.

This is not really an uncommon practice. I know many developers that do this very thing to avoid conflicts

Patch-set: 5