Update patch set 5
Patch Set 5:

> This needs serious fixing...  There are almost no cases where
 > someone should be altering authorship using "git commit --amend".
 > 
 > Manually cherry-picking using "git cherry-pick" preserves
 > authorship
 > 
 > Applying a patch grabbed from github using "git am" also preserves
 > authorship

Exactly why I brought up adding the "--author" (and "--date") flag for "git commit". I should have been more clear in my recommendation, as I meant for the entire reference to "git commit --amend" be removed.

If someone adds a change without using cherry-pick, adding the "--author=" flag when they commit the changes solves the problem of authorship not being "automagically" preserved.

Amending authorship to a commit should only be done if the committer forgot to add the --author flag when committing changes in the first place, or if they can't type.

Patch-set: 5