twrp-8.1: sync a shallow clone of frameworks folders
".repo" is no longer allowed in folder name in newer versions of repo,
so we now need to sync some frameworks folders and dependencies
for builds to run/finish successfully.

This adds ~1GB to the overall downloaded source, but there is no
workaround for this anymore since repo was updated.

Change-Id: Ica996ad5e9e8595868011c4246d479eac5777b48